requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kafka-python==2.0.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
oauthlib==3.1.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
py4j==0.10.9
pycparser==2.20
Pygments==2.8.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pyspark==3.0.1
python-dateutil==2.8.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
Send2Trash==1.5.0
six==1.15.0
terminado==0.9.2
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
tweepy==3.10.0
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
